Included Home Assistant support with MQTT Autodiscovery
You should be aware, this breaks pretty everything of your formerly used setup. I removed the whole part of your board and replaced it so it works with ESP32.

Also this will currently assume that you don't use a battery powered device. Maybe there should be a setting for that so we can switch this easily and support both modes.